rapidapi_version/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Web scraping
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0

# Data processing
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0